+++ bulldog.devices/src/main/java/org/bulldog/devices/pwmdriver/FrequencyLookupTable.java
@@ -2,7 +2,7 @@ package org.bulldog.devices.pwmdriver;
 
 import java.util.HashMap;
 
-public class FrequencyLookupTable extends HashMap<Float, Integer> {
+public class FrequencyLookupTable extends HashMap<Double, Integer> {
 
 	/**
 	 * 
@@ -10,7 +10,7 @@ public class FrequencyLookupTable extends HashMap<Float, Integer> {
 	private static final long serialVersionUID = 4953943440394262798L;
 
 	public FrequencyLookupTable() {
-		this.put(50.0f, 134);
+		this.put(50.0, 134);
 	}
 		
 }
