+++ bulldog.devices/src/main/java/org/bulldog/devices/servo/ServoListener.java
@@ -2,7 +2,7 @@ package org.bulldog.devices.servo;
 
 public interface ServoListener {
 
-	void angleChanged(Servo servo, float oldAngle, float newAngle);
-	void moveCompleted(Servo servo, float oldAngle, float newAngle);
+	void angleChanged(Servo servo, double oldAngle, double newAngle);
+	void moveCompleted(Servo servo, double oldAngle, double newAngle);
 	
 }
